refactor: 게임 실행 횟수 검증 클래스 디렉토리 변경
